+++ config/routes.rb
@@ -3,6 +3,11 @@ Rails.application.routes.draw do
   mount_avo
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
+  # Webhooks
+  namespace :webhooks do
+    post "whatsapp/incoming", to: "whatsapp#incoming"
+  end
+
   # Reveal health status on /up that returns 200 if the app boots with no exceptions, otherwise 500.
   # Can be used by load balancers and uptime monitors to verify that the app is live.
   get "up" => "rails/health#show", as: :rails_health_check